pyproject.toml
[tool.poetry]
name = "sentinel"
version = "0.1.0"
description = "Sentinel is an audit storage system"
authors = ["Oscar Carballal <oscar@lareira.digital>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
pytest = "*"
hypercorn = "*"
PyJWT = "*"
ujson = "*"
pytz = "*"
python-dotenv = "*"
immudb-py = "*"
uvloop = "*"

[tool.poetry.dev-dependencies]
ipython = "*"
ipdb = "*"
black = "*"
pytest-vcr = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
